#32 Fixing collection out of date
Refactoring elasticsearch

Switching match to term and filter to must, no need for fuzzy matching or scoring
