+++ src/app/api/score/score.py
@@ -1,15 +1,17 @@
 from uuid import UUID
 
+from elasticsearch_dsl import Q
 from elasticsearch_dsl.response import Response
 from fastapi import Path
 
 import app.core.constants
+from app.api.score.models import CollectionAttribute, LearningMaterialAttribute
+from app.elastic.dsl import afilter, amissing
 from app.elastic.elastic import (
     ResourceType,
-    aggs_collection_validation,
-    aggs_material_validation,
     query_collections,
     query_materials,
+    query_missing_material_license,
 )
 from app.elastic.search import Search
 
@@ -55,7 +57,7 @@ def score_search(noderef_id: UUID, resource_type: ResourceType) -> Search:
         query, aggs = query_collections, aggs_collection_validation
     elif resource_type is ResourceType.MATERIAL:
         query, aggs = query_materials, aggs_material_validation
-    s = Search().query(query(ancestor_id=noderef_id))
+    s = Search().base_filters().query(query(ancestor_id=noderef_id))
     for name, _agg in aggs.items():
         s.aggs.bucket(name, _agg)
     return s
@@ -82,3 +84,42 @@ def collection_id_param(
     collection_id: UUID = Path(..., examples=app.core.constants.COLLECTION_NAME_TO_ID),
 ) -> UUID:
     return collection_id
+
+
+def field_names_used_for_score_calculation(properties: dict) -> list[str]:
+    values = []
+    for value in properties.values():
+        value = list(list(value.to_dict().values())[0].values())[0]
+        if isinstance(value, dict):
+            value = list(value.keys())[0]
+
+        value.replace(".keyword", "")
+        if value != "should":
+            values += [value]
+    return values
+
+
+aggs_material_validation = {
+    "missing_title": amissing(qfield=LearningMaterialAttribute.TITLE),
+    "missing_keywords": amissing(qfield=LearningMaterialAttribute.KEYWORDS),
+    "missing_subjects": amissing(qfield=LearningMaterialAttribute.SUBJECTS),
+    "missing_description": amissing(qfield=LearningMaterialAttribute.DESCRIPTION),
+    "missing_license": afilter(query=query_missing_material_license()),
+    "missing_edu_context": amissing(qfield=LearningMaterialAttribute.EDU_CONTEXT),
+    "missing_ads_qualifier": amissing(qfield=LearningMaterialAttribute.CONTAINS_ADS),
+    "missing_material_type": amissing(
+        qfield=LearningMaterialAttribute.LEARNINGRESOURCE_TYPE
+    ),
+    "missing_object_type": amissing(qfield=LearningMaterialAttribute.OBJECT_TYPE),
+}
+aggs_collection_validation = {
+    "missing_title": amissing(qfield=CollectionAttribute.TITLE),
+    "short_title": afilter(Q("range", char_count_title={"gt": 0, "lt": 5})),
+    "missing_keywords": amissing(qfield=CollectionAttribute.KEYWORDS),
+    "few_keywords": afilter(Q("range", token_count_keywords={"gt": 0, "lt": 3})),
+    "missing_description": amissing(qfield=CollectionAttribute.DESCRIPTION),
+    "short_description": afilter(
+        Q("range", char_count_description={"gt": 0, "lt": 30})
+    ),
+    "missing_edu_context": amissing(qfield=CollectionAttribute.EDU_CONTEXT),
+}
