+++ src/app/elastic/filters.py
@@ -1,0 +1,8 @@
+from app.elastic.dsl import qterm
+from app.models import ElasticResourceAttribute
+
+base_filter = [
+    qterm(qfield=ElasticResourceAttribute.PERMISSION_READ, value="GROUP_EVERYONE"),
+    qterm(qfield=ElasticResourceAttribute.EDU_METADATASET, value="mds_oeh"),
+    qterm(qfield=ElasticResourceAttribute.PROTOCOL, value="workspace"),
+]
